Refactor: improve the loading of the data at P10.
Now it is done sequentially to avoid running out of memory.

The parameter is_series as been introduced: for several images with same motor positions, images are summed by default.

The .fio file is not reliable, and nothing indicates if the scan is composed of series or not inside it.